long_memory/requirements.txt
flask
pymysql-pooling
pinecone-client
sentence_transformers
pymysql
openai